+++ XuanYu.Core/Gizmo/MoveGizmoAxis.cs
@@ -1,0 +1,8 @@
+namespace XuanYu.Core.Gizmo;
+
+public enum MoveGizmoAxis
+{
+    X,
+    Y,
+    Z
+}